refactor(tui): colocate Screen::Pins into tui/screens/pins.rs
Fifth screen migration for #287 (Phase 2). Moved into the new
tui/screens/pins.rs: handle_key_pins, pins_guard (the #288 shared guard,
Pins-exclusive), pin_row_label (only ever called from build_pins_vm and
tests — same "exclusive helper physically in the wrong file" pattern as
diff_footer/adjust_config_field), build_pins_vm, render_pins_vm,
pins_palette_items.

diff_pair_previewable (keys.rs) is shared by pins_guard and list_guard (List
hasn't migrated yet), so it stays in keys.rs, widened to pub(crate).
apply_filter_edit/FilterKey (keys.rs) are shared by Pins/Gists/List's inline
filter-editing, also widened to pub(crate) rather than moved.

render_pins_vm's filtering branch still reads state.pins().expect(...)
directly for the live TextInput (PinsVm.filter_query is a plain String,
which input_line can't use for cursor position) — the same pre-existing gap
introduced or fixed here; carried over verbatim as part of the pure move.

Pure move — bodies and existing tests unchanged.

Part of #287
