+++ src/tui/palette.rs
@@ -1,5 +1,5 @@
 use super::{
-    keys::{detail_guard, gists_guard, list_guard, pins_guard, point_in, revisions_guard},
+    keys::{detail_guard, gists_guard, list_guard, point_in, revisions_guard},
     *,
 };
 use crossterm::event::{KeyCode, KeyModifiers};
@@ -321,7 +321,7 @@ fn cross_items() -> Vec<PaletteItem> {
 fn build_palette_items(state: &AppState, screen: &Screen, mode: PaletteMode) -> Vec<PaletteItem> {
     let mut items = match screen {
         Screen::List => list_palette_items(state),
-        Screen::Pins(_) => pins_palette_items(state),
+        Screen::Pins(_) => super::screens::pins::pins_palette_items(state),
         Screen::Gists(_) => gists_palette_items(state),
         Screen::GistDetail(_) => detail_palette_items(state),
         Screen::Revisions(_) => revisions_palette_items(state),
@@ -426,26 +426,6 @@ fn list_palette_items(state: &AppState) -> Vec<PaletteItem> {
         key_item("t", "Toggle description / id", KeyCode::Char('t'), true),
         key_item("v", "Cycle gist visibility", KeyCode::Char('v'), true),
         key_item("s", "Cycle pane sort", KeyCode::Char('s'), true),
-        key_item("?", "Help", KeyCode::Char('?'), true),
-    ]
-}
-
-fn pins_palette_items(state: &AppState) -> Vec<PaletteItem> {
-    let g = |code| pins_guard(state, code);
-    vec![
-        key_item(
-            "Enter",
-            "Diff pinned pair",
-            KeyCode::Enter,
-            g(KeyCode::Enter),
-        ),
-        key_item("s", "Smart-sync", KeyCode::Char('s'), g(KeyCode::Char('s'))),
-        key_item("u", "Force push", KeyCode::Char('u'), g(KeyCode::Char('u'))),
-        key_item("d", "Force pull", KeyCode::Char('d'), g(KeyCode::Char('d'))),
-        key_item("x", "Unpin pair", KeyCode::Char('x'), g(KeyCode::Char('x'))),
-        key_item("/", "Filter pins", KeyCode::Char('/'), true),
-        key_item("o", "Cycle sort", KeyCode::Char('o'), true),
-        key_item("q", "Back to list", KeyCode::Char('q'), true),
         key_item("?", "Help", KeyCode::Char('?'), true),
     ]
 }
